Se añadió Envio y MetodoPago, se cambió Integer por Long en los Ids

+++ src/main/java/com/example/elarayax/naves/controller/CategoriaController.java
@@ -31,7 +31,7 @@ public class CategoriaController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Categoria> getCategoriaById(@PathVariable Integer id) {
+    public ResponseEntity<Categoria> getCategoriaById(@PathVariable Long id) {
         Categoria categoria = categoriaService.findById(id);
         if (categoria == null) {
             return ResponseEntity.notFound().build();
